+++ examples/upload-wait-download.js
@@ -6,23 +6,23 @@ const { getData, getStatus, uploadPdf } = require("..");
 
 function printResults(data) {
     console.log(
-        `Got results for file ${data.fileName} with ${data.pages.length} pages`
+        `Got results for file ${data.fileName} with ${data.pages.length} pages`,
     );
 
     let pageNum = 0;
     for (const page of data.pages) {
         const points = page.groups.reduce(
             (acc, group) => acc + group.points.length,
-            0
+            0,
         );
         const paths = page.groups.reduce(
             (acc, group) => acc + group.paths.length,
-            0
+            0,
         );
         console.log(
             `Page ${++pageNum} has ${
                 page.groups.length
-            } groups with ${points} points and ${paths} paths`
+            } groups with ${points} points and ${paths} paths`,
         );
         // console.log(`Example path:\n${JSON.stringify(page.groups[0].paths[0], undefined, 4)}`);
         // console.log("Page data", page);
@@ -51,14 +51,14 @@ function waitAndDownload(settings, token) {
 function main() {
     if (!process.env.API_KEY) {
         console.error(
-            "Please set the 'API_KEY' environment variable to use the command line tester."
+            "Please set the 'API_KEY' environment variable to use the command line tester.",
         );
         return;
     }
 
     if (process.argv.length < 3) {
         console.error(
-            "Please give the filename to upload to use the command line tester."
+            "Please give the filename to upload to use the command line tester.",
         );
         return;
     }